+++ test/test_util.py
@@ -1,3 +1,4 @@
+# coding: utf-8
 import hashlib
 import warnings
 import logging
@@ -42,6 +43,8 @@ from urllib3.util import is_fp_closed, ssl_
 from urllib3.packages import six
 
 from . import clear_warnings
+
+from test import onlyPy3, onlyPy2
 
 # This number represents a time in seconds, it doesn't mean anything in
 # isolation. Setting to a high-ish value to avoid conflicts with the smaller
@@ -152,12 +155,15 @@ class TestUtil(object):
         # Path/query/fragment
         ('', Url()),
         ('/', Url(path='/')),
+        ('/abc/../def', Url(path="/abc/../def")),
         ('#?/!google.com/?foo#bar', Url(path='', fragment='?/!google.com/?foo#bar')),
         ('/foo', Url(path='/foo')),
         ('/foo?bar=baz', Url(path='/foo', query='bar=baz')),
         ('/foo?bar=baz#banana?apple/orange', Url(path='/foo',
                                                  query='bar=baz',
                                                  fragment='banana?apple/orange')),
+        ('/redirect?target=http://localhost:61020/', Url(path='redirect',
+                                                         query='target=http://localhost:61020/')),
 
         # Port
         ('http://google.com/', Url('http', host='google.com', path='/')),
@@ -171,7 +177,16 @@ class TestUtil(object):
                                               auth='foo:bar@baz',
                                               host='localhost',
                                               path='/')),
-        ('http://@', Url('http', host=None, auth=''))
+
+        # Unicode type (Python 2.x)
+        (u'http://foo:bar@localhost/', Url(u'http',
+                                           auth=u'foo:bar',
+                                           host=u'localhost',
+                                           path=u'/')),
+        ('http://foo:bar@localhost/', Url('http',
+                                          auth='foo:bar',
+                                          host='localhost',
+                                          path='/')),
     ]
 
     non_round_tripping_parse_url_host_map = [
@@ -197,8 +212,12 @@ class TestUtil(object):
         assert url == expected_url.url
 
     def test_parse_url_invalid_IPv6(self):
-        with pytest.raises(ValueError):
+        with pytest.raises(LocationParseError):
             parse_url('[::1')
+
+    def test_parse_url_negative_port(self):
+        with pytest.raises(LocationParseError):
+            parse_url("https://www.google.com:-80/")
 
     def test_Url_str(self):
         U = Url('http', host='google.com')
@@ -231,6 +250,54 @@ class TestUtil(object):
     @pytest.mark.parametrize('url, expected_netloc', url_netloc_map)
     def test_netloc(self, url, expected_netloc):
         assert parse_url(url).netloc == expected_netloc
+
+    url_vulnerabilities = [
+        # urlparse doesn't follow RFC 3986 Section 3.2
+        ("http://google.com#@evil.com/", Url("http",
+                                             host="google.com",
+                                             path="",
+                                             fragment="@evil.com/")),
+
+        # CVE-2016-5699
+        ("http://127.0.0.1%0d%0aConnection%3a%20keep-alive",
+         Url("http", host="127.0.0.1%0d%0aConnection%3a%20keep-alive")),
+
+        # NodeJS unicode -> double dot
+        (u"http://google.com/\uff2e\uff2e/abc", Url("http",
+                                                    host="google.com",
+                                                    path='/%ef%bc%ae%ef%bc%ae/abc'))
+    ]
+
+    @pytest.mark.parametrize("url, expected_url", url_vulnerabilities)
+    def test_url_vulnerabilities(self, url, expected_url):
+        if expected_url is False:
+            with pytest.raises(LocationParseError):
+                parse_url(url)
+        else:
+            assert parse_url(url) == expected_url
+
+    @onlyPy2
+    def test_parse_url_bytes_to_str_python_2(self):
+        url = parse_url(b"https://www.google.com/")
+        assert url == Url('https', host='www.google.com', path='/')
+
+        assert isinstance(url.scheme, str)
+        assert isinstance(url.host, str)
+        assert isinstance(url.path, str)
+
+    @onlyPy2
+    def test_parse_url_unicode_python_2(self):
+        url = parse_url(u"https://www.google.com/")
+        assert url == Url(u'https', host=u'www.google.com', path=u'/')
+
+        assert isinstance(url.scheme, six.text_type)
+        assert isinstance(url.host, six.text_type)
+        assert isinstance(url.path, six.text_type)
+
+    @onlyPy3
+    def test_parse_url_bytes_type_error_python_3(self):
+        with pytest.raises(TypeError):
+            parse_url(b"https://www.google.com/")
 
     @pytest.mark.parametrize('kwargs, expected', [
         ({'accept_encoding': True},
@@ -300,6 +367,7 @@ class TestUtil(object):
         (('abcd', ''),   ('abcd', '', None)),
         (('abcd', 'a'),  ('', 'bcd', 'a')),
         (('abcd', 'ab'), ('', 'bcd', 'a')),
+        (('abcd', 'eb'), ('a', 'cd', 'b')),
     ])
     def test_split_first(self, input, expected):
         output = split_first(*input)